+++ src/main/java/dk/kb/datahandler/oai/OaiResponseFilter.java
@@ -14,12 +14,14 @@
  */
 package dk.kb.datahandler.oai;
 
-import dk.kb.storage.invoker.v1.ApiException;
 import dk.kb.storage.model.v1.DsRecordDto;
 import dk.kb.storage.model.v1.RecordTypeDto;
 import dk.kb.storage.util.DsStorageClient;
+import dk.kb.util.webservice.exception.ServiceException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.xml.sax.helpers.DefaultHandler;
 
 
 /**
@@ -47,7 +49,7 @@ public class OaiResponseFilter {
      * sets the type to {@link RecordTypeDto#DELIVERABLEUNIT}.
      * @param response      OAI-PMH response containing records.
      */
-    public void addToStorage(OaiResponse response) throws ApiException {
+    public void addToStorage(OaiResponse response) throws ServiceException {
         for (OaiRecord oaiRecord: response.getRecords()) {
             addToStorage(oaiRecord);
             processed++;
@@ -59,13 +61,23 @@ public class OaiResponseFilter {
      * sets the type to {@link RecordTypeDto#DELIVERABLEUNIT}.
      * @param oaiRecord     a record from an OAI-PMH response
      */
-    public void addToStorage(OaiRecord oaiRecord) throws ApiException {
-        String origin = getOrigin(oaiRecord, datasource);
+    public void addToStorage(OaiRecord oaiRecord) throws ServiceException {
+        String origin = getOrigin(oaiRecord, datasource, null);
+        addToStorage(oaiRecord, origin);
+    }
+
+    /**
+     * Add record from an OAI-PMH harvest to ds-storage. The public implementation does not resolve parent and
+     * sets the type to {@link RecordTypeDto#DELIVERABLEUNIT}.
+     * @param oaiRecord     a record from an OAI-PMH response
+     * @param origin        the origin, which the record is added to in ds-storage.
+     */
+    public void addToStorage(OaiRecord oaiRecord, String origin) throws ServiceException {
         String storageId = origin + ":" + oaiRecord.getId();
         if (oaiRecord.isDeleted()) {
             storage.markRecordForDelete(storageId);
         } else if (origin.isEmpty()){
-           log.warn("OAI Record with ID: '{}', has empty origin, it is not added to DS-Storage.", oaiRecord.getId());
+            log.warn("OAI Record with ID: '{}', has empty origin, it is not added to DS-Storage.", oaiRecord.getId());
         }
         else {
             String parentID = getParentID(oaiRecord, origin);
@@ -79,9 +91,10 @@ public class OaiResponseFilter {
      * public behaviour is to return the {@code datasource}.
      * @param oaiRecord     a record from an OAI-PMH response
      * @param datasource    where the harvested OAI-PMH record is extracted from.
+     * @param recordHandler used to parse the data from the osiRecord if needed.
      * @return {@code origin} for the given {@code oaiRecord}.
      */
-    public String getOrigin(OaiRecord oaiRecord, String datasource) {
+    public String getOrigin(OaiRecord oaiRecord, String datasource, DefaultHandler recordHandler) {
         return datasource;
     }
 
@@ -119,7 +132,7 @@ public class OaiResponseFilter {
      * @param origin    the origin for the record.
      */
     private void addOrUpdateRecord(OaiRecord oaiRecord, String storageId, String parentID, String origin)
-            throws ApiException {
+            throws ServiceException {
         DsRecordDto dsRecord = new DsRecordDto();
         dsRecord.setId(storageId);
         dsRecord.setOrigin(origin);