+++ src/main/java/dk/kb/datahandler/oai/OaiResponseFilterDrArchive.java
@@ -1,13 +1,21 @@
 package dk.kb.datahandler.oai;
 
 import dk.kb.datahandler.enrichment.DataEnricher;
-import dk.kb.storage.invoker.v1.ApiException;
+import dk.kb.datahandler.util.PreservicaOaiRecordHandler;
 import dk.kb.storage.util.DsStorageClient;
+import dk.kb.util.webservice.exception.InternalServiceException;
+import dk.kb.util.webservice.exception.ServiceException;
+
 import org.eclipse.jetty.util.StringUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.xml.sax.SAXException;
 
-import java.util.regex.Matcher;
+import javax.xml.parsers.SAXParser;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.regex.Pattern;
 
 public class OaiResponseFilterDrArchive extends OaiResponseFilterPreservicaSeven{
@@ -16,8 +24,6 @@ public class OaiResponseFilterDrArchive extends OaiResponseFilterPreservicaSeven
 
     private String fragmentServiceUrl = null;
 
-    protected static final Pattern DR_PATTERN = Pattern.compile(
-            ">(?i:dr)[^<]*</(?:\\w+:)?publisher>");
 
     /**
      * @param datasource source for records. Default implementation uses this for {@code origin}.
@@ -40,9 +46,10 @@ public class OaiResponseFilterDrArchive extends OaiResponseFilterPreservicaSeven
      * @param response      OAI-PMH response containing preservica records.
      */
     @Override
-    public void addToStorage(OaiResponse response) throws ApiException {
+    public void addToStorage(OaiResponse response) throws ServiceException {
+        SAXParser saxParser = getSaxParser();
+
         for (OaiRecord oaiRecord: response.getRecords()) {
-            String xml = oaiRecord.getMetadata();
             String recordId = oaiRecord.getId();
             // Preservica StructuralObjects are ignored as they are only used as folders in the GUI.
             if (recordId.contains("oai:so")){
@@ -50,9 +57,16 @@ public class OaiResponseFilterDrArchive extends OaiResponseFilterPreservicaSeven
                 continue;
             }
 
+            PreservicaOaiRecordHandler handler = new PreservicaOaiRecordHandler();
+            try {
+                InputStream inputXml = new ByteArrayInputStream(oaiRecord.getMetadata().getBytes(StandardCharsets.UTF_8));
+                saxParser.parse(inputXml, handler);
+            } catch (SAXException | IOException e) {
+                throw new InternalServiceException(e);
+            }
+
             // Filter out material that are not send on DR channels
-            Matcher drMatcher = DR_PATTERN.matcher(xml);
-            if (!drMatcher.find()){
+            if (!handler.recordIsDr){
                 processed++;
                 nonDrRecords++;
                 // Periodically logging of how many records have been filtered out.
@@ -64,37 +78,25 @@ public class OaiResponseFilterDrArchive extends OaiResponseFilterPreservicaSeven
             }
 
             // InformationObjects from preservica 7 need to have the PBCore metadata tag.
-            Matcher metadataMatcher = METADATA_PATTERN.matcher(xml);
-            if ((recordId.contains("oai:io")) && !metadataMatcher.find()) {
-                processed++;
-                emptyMetadataRecords ++;
-                log.warn("OAI-PMH record '{}' does not contain PBCore metadata and is therefore not added to storage. " +
-                                "'{}' empty records have been found and '{}' records have been processed in total.",
-                        recordId, emptyMetadataRecords, processed);
+            if (!informationObjectContainsPbcoreBoolean(handler, recordId)){
                 continue;
             }
 
-            Matcher transcodingDoneMatcher = TRANSCODING_PATTERN.matcher(xml);
-            if (!transcodingDoneMatcher.find()) {
-                processed++;
-                transCodingNotDoneRecords++;
-                log.debug("OAI-PMH record '{}' transcoding status not done. Record skipped",recordId);
-                if (transCodingNotDoneRecords % 1000 == 0) {
-                    log.info("'{}' records transcoding status not done filtered away. '{}' records have been processed.",
-                            transCodingNotDoneRecords, processed);
-                }
+            if (!transcodingStatusIsDoneBoolean(handler, recordId)){
                 continue;
             }
+
+            String origin = getOrigin(oaiRecord, datasource, handler);
 
             // Enrich record
-            if (!StringUtil.isEmpty(fragmentServiceUrl) && "ds.tv".equals(getOrigin(oaiRecord,datasource))) {
-                    DataEnricher.apply(fragmentServiceUrl,oaiRecord);
+            if (!StringUtil.isEmpty(fragmentServiceUrl) && "ds.tv".equals(origin)) {
+                DataEnricher.apply(fragmentServiceUrl,oaiRecord);
             }
 
             try {
-                addToStorage(oaiRecord);
+                addToStorage(oaiRecord, origin);
                 processed++;
-            } catch (ApiException e){
+            } catch (ServiceException e){
                 log.warn("DsStorage threw an exception when adding OAI record from Preservica 7 to storage.");
                 throw e;
             }