+++ src/main/java/dk/kb/datahandler/oai/OaiResponseFilterPreservicaSeven.java
@@ -1,14 +1,27 @@
 package dk.kb.datahandler.oai;
 
-import dk.kb.storage.invoker.v1.ApiException;
-import dk.kb.storage.model.v1.DsRecordDto;
-import dk.kb.storage.model.v1.RecordTypeDto;
-import dk.kb.storage.util.DsStorageClient;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import dk.kb.datahandler.util.PreservicaOaiRecordHandler;
+import dk.kb.util.webservice.exception.InternalServiceException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import dk.kb.storage.model.v1.DsRecordDto;
+import dk.kb.storage.model.v1.RecordTypeDto;
+import dk.kb.storage.util.DsStorageClient;
+import dk.kb.util.webservice.exception.ServiceException;
+import org.xml.sax.SAXException;
+import org.xml.sax.helpers.DefaultHandler;
+
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.parsers.SAXParser;
+import javax.xml.parsers.SAXParserFactory;
 
 /**
  * Filtering and delivery of Preservica OAI records from Preservica 7. Generates {@code datasource} prefixed IDs,
@@ -16,36 +29,7 @@ import java.util.regex.Pattern;
  */
 public class OaiResponseFilterPreservicaSeven extends OaiResponseFilter{
     private static final Logger log = LoggerFactory.getLogger(OaiResponseFilterPreservicaSeven.class);
-
-    /**
-     * Pattern for determining if a InformationObject from Preservica 7
-     * contains metadata about a radio resource.
-     * The regex handles the fact that namespace prefixes are arbitrarily defined.
-     */
-    protected static final Pattern RADIO_PATTERN = Pattern.compile(
-            ">Sound</(?:\\w+:)?formatMediaType", Pattern.CASE_INSENSITIVE);
-
-    /**
-     * Pattern for determining if a InformationObject from Preservica 7
-     * contains metadata about a television resource.
-     * The regex handles the fact that namespace prefixes are arbitrarily defined.
-     */
-    protected static final Pattern TV_PATTERN = Pattern.compile(
-            ">Moving\\sImage</(?:\\w+:)?formatMediaType", Pattern.CASE_INSENSITIVE);
-
-    /**
-     * Pattern used to check that records does in fact contain PBCore metadata.
-     */
-    static final Pattern METADATA_PATTERN = Pattern.compile(
-            "<Metadata\\s+schemaUri=\"http://www\\.pbcore\\.org/PBCore/PBCoreNamespace\\.html\">");
-
-    /**
-     * Pattern used for checking the transcoding status of a record. Only records with transcoding status done should be added.
-     */
-
-    static final Pattern TRANSCODING_PATTERN = Pattern.compile(
-            "<radiotvTranscodingStatus:radiotvTranscodingStatus(?s).*<transcodingStatus>done</transcodingStatus>(?s).*</radiotvTranscodingStatus:radiotvTranscodingStatus>"
-    );
+    static final SAXParserFactory factory = SAXParserFactory.newInstance();
 
     protected int emptyMetadataRecords = 0;
 
@@ -66,64 +50,110 @@ public class OaiResponseFilterPreservicaSeven extends OaiResponseFilter{
      * @param response      OAI-PMH response containing preservica records.
      */
     @Override
-    public void addToStorage(OaiResponse response) throws ApiException {
+    public void addToStorage(OaiResponse response) throws ServiceException {
+        SAXParser saxParser = getSaxParser();
+
         for (OaiRecord oaiRecord: response.getRecords()) {
-            String xml = oaiRecord.getMetadata();
-            String recordId = oaiRecord.getId();
-            // Preservica StructuralObjects are ignored as they are only used as folders in the GUI.
-            if (recordId.contains("oai:so")){
-                //log.debug("Skipped Structural object with id: '{}'", recordId);
-                continue;
-            }
-            // InformationObjects from preservcia 6/7 need to have the PBCore metadata tag.
-            Matcher metadataMatcher = METADATA_PATTERN.matcher(xml);
-            if ((recordId.contains("oai:io")) && !metadataMatcher.find()) {
-                processed++;
-                emptyMetadataRecords ++;
-                log.warn("OAI-PMH record '{}' does not contain PBCore metadata and is therefore not added to storage. " +
-                                "'{}' empty records have been found and '{}' records have been processed in total.",
-                        recordId, emptyMetadataRecords, processed);
-                continue;
-            }
-
-            Matcher transcodingDoneMatcher = TRANSCODING_PATTERN.matcher(xml);
-            if (!transcodingDoneMatcher.find()) {
-                processed++;
-                transCodingNotDoneRecords++;
-                log.debug("OAI-PMH record '{}' transcoding status not done. Record skipped",recordId);
-                if (transCodingNotDoneRecords % 1000 == 0) {
-                    log.info("'{}' records transcoding status not done filtered away. '{}' records have been processed.",
-                            transCodingNotDoneRecords, processed);
-                }
-                continue;
-            }
-
             try {
-                addToStorage(oaiRecord);
+                PreservicaOaiRecordHandler handler = new PreservicaOaiRecordHandler();
+
+                InputStream inputXml = new ByteArrayInputStream(oaiRecord.getMetadata().getBytes(StandardCharsets.UTF_8));
+                saxParser.parse(inputXml, handler);
+
+                String recordId = oaiRecord.getId();
+                // Preservica StructuralObjects are ignored as they are only used as folders in the GUI.
+                if (recordId.contains("oai:so")){
+                    log.debug("Skipped Structural object with id: '{}'", recordId);
+                    continue;
+                }
+
+                // InformationObjects from preservcia 6/7 need to have the PBCore metadata tag.
+                if (!informationObjectContainsPbcoreBoolean(handler, recordId)){
+                    continue;
+                }
+
+                if (!transcodingStatusIsDoneBoolean(handler, recordId)){
+                    continue;
+                }
+
+                String origin = getOrigin(oaiRecord, datasource, handler);
+
+                addToStorage(oaiRecord, origin);
                 processed++;
-            } catch (ApiException e){
+            } catch (ServiceException e){
                 log.warn("DsStorage threw an exception when adding OAI record from Preservica 7 to storage.");
                 throw e;
+            } catch (IOException | SAXException e) {
+                throw new InternalServiceException("An error occurred when parsing XML with SAX:", e);
             }
         }
     }
 
+    /**
+     * Checks if the transcoding status is complete for the given XML string.
+     *
+     * <p>This method uses a regex pattern to determine if the transcoding status is marked as done in the XML.
+     * If the status is not found, it increments the counters for processed records and those with
+     * transcoding not done, logs a debug message, and, for every 1000 records filtered, logs an info message.
+     * The method returns false if the transcoding status is not complete; otherwise, it returns true.</p>
+     *
+     * @param recordHandler The XML string containing the transcoding status to be checked.
+     * @param recordId The unique identifier for the record being checked.
+     * @return {@code true} if the transcoding status is complete; {@code false} if the status is not done.
+     */
+    boolean transcodingStatusIsDoneBoolean(PreservicaOaiRecordHandler recordHandler, String recordId) {
+        if (!recordHandler.recordIsTranscoded){
+            processed++;
+            transCodingNotDoneRecords++;
+            log.debug("OAI-PMH record '{}' transcoding status not done. Record skipped", recordId);
+            if (transCodingNotDoneRecords % 1000 == 0) {
+                log.info("'{}' records transcoding status not done filtered away. '{}' records have been processed.",
+                        transCodingNotDoneRecords, processed);
+            }
+            return false;
+        }
+
+        return true;
+    }
+
+
+    /**
+     * Checks if the given XML string contains PBCore metadata and that the record is an InformationObject.
+     *
+     * @param recordHandler containing a boolean describing if the record in hand contains PBCore metadata.
+     * @param recordId The unique identifier for the record, which indicates if the record is an InformationObject.
+     * @return {@code true} if the XML contains PBCore metadata or if the record ID does not indicate
+     *         an OAI-PMH record; {@code false} if the record ID indicates an OAI-PMH record
+     *         and PBCore metadata is not found.
+     */
+    boolean informationObjectContainsPbcoreBoolean(PreservicaOaiRecordHandler recordHandler, String recordId) {
+        if (recordId.contains("oai:io") && !recordHandler.recordHasMetadata){
+            processed++;
+            emptyMetadataRecords ++;
+            log.warn("OAI-PMH record '{}' does not contain PBCore metadata and is therefore not added to storage. " +
+                            "'{}' empty records have been found and '{}' records have been processed in total.",
+                    recordId, emptyMetadataRecords, processed);
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Get DS origin from OaiRecordHandler which have streamed the content from the incoming OAI record.
+     */
     @Override
-    public String getOrigin(OaiRecord oaiRecord, String datasource) {
-        String xml = oaiRecord.getMetadata();
+    public String getOrigin(OaiRecord oaiRecord, String datasource, DefaultHandler recordHandler) {
+        PreservicaOaiRecordHandler preservicaRecordHandler = (PreservicaOaiRecordHandler) recordHandler;
 
-        Matcher radioDeliverableunitMatcher = RADIO_PATTERN.matcher(xml);
-        Matcher tvDeliverableUnitMatcher = TV_PATTERN.matcher(xml);
-
-        if (radioDeliverableunitMatcher.find()){
-            return "ds.radio";
-        } else if (tvDeliverableUnitMatcher.find()){
-            return "ds.tv";
-        } else {
-            log.warn("No specific origin has been extracted for preservica record '{}'.",
-                    oaiRecord.getId());
-            // Not quite sure what we should do in the case where nothing gets matched.
-            return "";
+        switch (preservicaRecordHandler.getRecordType()) {
+            case TV:
+                return "ds.tv";
+            case RADIO:
+                return "ds.radio";
+            case UNKNOWN:
+                return "";
+            default:
+                throw new InternalServiceException("Unknown record type: " + preservicaRecordHandler.getRecordType());
         }
     }
 
@@ -142,5 +172,13 @@ public class OaiResponseFilterPreservicaSeven extends OaiResponseFilter{
 
         log.debug("Unable to derive record type for id '{}' from datasource '{}'. Falling back to '{}'", storageId, datasource, RecordTypeDto.DELIVERABLEUNIT);
         return RecordTypeDto.DELIVERABLEUNIT;
+    }
+
+    static SAXParser getSaxParser() {
+        try {
+            return factory.newSAXParser();
+        } catch (ParserConfigurationException | SAXException e) {
+            throw new InternalServiceException("An error occurred when constructing SAXParser: ", e);
+        }
     }
 }